requirements.txt
appdirs==1.4.3
beautifulsoup4==4.5.3
bs4==0.0.1
Cython==0.25.2
Django==1.11
image==1.5.5
olefile==0.44
packaging==16.8
Pillow==4.1.1
pycurl==7.43.0
pyparsing==2.2.0
pytesseract==0.1.6
pytz==2017.2
requests==2.13.0
selenium==3.4.1
six==1.10.0
